backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
supabase==2.3.4
chromadb==0.4.22
sentence-transformers==2.3.1
groq==0.4.2
langchain==0.1.6
langchain-community==0.0.19
pymupdf==1.23.8
python-docx==1.1.0
python-jose[cryptography]==3.3.0
cachetools==5.3.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
structlog==24.1.0
httpx==0.26.0
